Resilience_Agent/requirements.txt
strands-agents-tools
strands-agents
mcp
fastapi
pydantic
requests
colorama
uvicorn
rich
agentcore
aws-opentelemetry-distro>=0.10.0
boto3